Convert to unsigned long long.  Set/Get options need to unsigned long long.  Read in options.  Other fixes.

+++ src/libsrcml/libsrcml.cpp
@@ -356,7 +356,7 @@ int srcml_set_version(const char* version) {
  *
  * @returns Return SRCML_STATUS_OK success and SRCML_STATUS_ERROR on failure.
  */
-int srcml_set_all_options(int option) {
+int srcml_set_all_options(unsigned long long option) {
 
   return srcml_archive_set_all_options(&global_archive, option);
 
@@ -370,7 +370,7 @@ int srcml_set_all_options(int option) {
  *
  * @returns Return SRCML_STATUS_OK success and SRCML_STATUS_ERROR on failure.
  */
-int srcml_set_option(int option) {
+int srcml_set_option(unsigned long long option) {
 
   return srcml_archive_set_option(&global_archive, option);
 
@@ -384,7 +384,7 @@ int srcml_set_option(int option) {
  *
  * @returns Return SRCML_STATUS_OK success and SRCML_STATUS_ERROR on failure.
  */
-int srcml_clear_option(int option) {
+int srcml_clear_option(unsigned long long option) {
 
   return srcml_archive_clear_option(&global_archive, option);
 
@@ -502,7 +502,7 @@ const char* srcml_get_version() {
  *
  * @returns Get the currently set options on success and NULL on failure.
  */
-int srcml_get_options() {
+unsigned long long srcml_get_options() {
 
   return srcml_archive_get_options(&global_archive);
